# public/docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "PolariixApi",
        "description": "Documentation complète de l'API - Tous les endpoints sont accessibles ici",
        "contact": {
            "name": "Support API",
            "email": "contact@votreapi.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/",
            "description": "Serveur API"
        }
    ],
    "paths": {
...
        "/api/v1/blogs": {
            "get": {
                "tags": [
                    "Blogs"
                ],
                "summary": "Liste de tous les articles de blog",
                "description": "Récupère la liste paginée de tous les articles de blog avec leurs auteurs et commentaires",
                "operationId": "49ee5f5b11b0fe9409399029dac30f30",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Numéro de page",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 1,
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Liste récupérée avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/SuccessResponse"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Erreur serveur",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Blogs"
                ],
                "summary": "Créer un nouvel article de blog",
                "description": "Crée un nouvel article de blog avec upload d'image optionnel. L'utilisateur doit être authentifié.",
                "operationId": "678ca615c913b934ed5db814351a737d",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "title",
                                    "content",
                                    "category"
                                ],
                                "properties": {
                                    "title": {
                                        "description": "Titre de l'article",
                                        "type": "string",
                                        "maxLength": 255,
                                        "example": "Comment devenir développeur en 2024"
                                    },
                                    "content": {
                                        "description": "Contenu de l'article en HTML ou texte",
                                        "type": "string",
                                        "example": "Contenu complet de l'article..."
                                    },
                                    "category": {
                                        "description": "Catégorie de l'article",
                                        "type": "string",
                                        "enum": [
                                            "Action",
                                            "Développement personnel",
                                            "Technologie",
                                            "Business",
                                            "Santé",
                                            "Lifestyle",
                                            "Éducation",
                                            "Divertissement",
                                            "Culture",
                                            "Voyage"
                                        ],
                                        "example": "Technologie"
                                    },
                                    "image": {
                                        "description": "Image de couverture (JPG, JPEG, PNG - Max 2MB)",
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Article créé avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/SuccessResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Non authentifié",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Erreur de validation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ValidationError"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Erreur serveur",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/api/v1/blogs/slug/{slug}": {
            "get": {
                "tags": [
                    "Blogs"
                ],
                "summary": "Consulter un article via son slug (URL publique SEO‑friendly)",
                "description": "Récupère un article par son slug URL‑friendly. À utiliser côté front pour des URLs lisibles et pour éviter d'exposer des IDs incrémentaux (anti‑énumération).",
                "operationId": "047c3fe38e4f0280abe6fe4a034d8e96",
                "parameters": [
                    {
                        "name": "slug",
                        "in": "path",
                        "description": "Slug de l'article (URL-friendly)",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "comment-devenir-developpeur-abc123"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Article trouvé",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/SuccessResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Article non trouvé",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Blogs"
                ],
                "summary": "Mettre à jour un article via son slug (référence URL‑friendly)",
                "description": "Met à jour un article en l'identifiant par son slug public, pratique quand seul l'URL publique est connue côté client.",
                "operationId": "fb6661277e06e5765ef71fd5337c52a1",
                "parameters": [
                    {
                        "name": "slug",
                        "in": "path",
                        "description": "Slug de l'article",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "comment-devenir-developpeur-abc123"
                        }
                    }
                ],
                "requestBody": {
                    "required": false,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "type": "string"
                                    },
                                    "content": {
                                        "type": "string"
                                    },
                                    "category": {
                                        "type": "string"
                                    },
                                    "image": {
                                        "type": "string",
                                        "format": "binary"
                                    },
                                    "_method": {
                                        "type": "string",
                                        "example": "PUT"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Article modifié avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/SuccessResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Blogs"
                ],
                "summary": "Supprimer un article via son slug (URL publique)",
                "description": "Supprime un article en le ciblant via son slug public, sans exposer l'ID interne.",
                "operationId": "566618ace26bec65b35e2b207477ad9a",
                "parameters": [
                    {
                        "name": "slug",
                        "in": "path",
                        "description": "Slug de l'article",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "comment-devenir-developpeur-abc123"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Article supprimé avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/SuccessResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "SuccessResponse": {
                "title": "Réponse de succès standard",
                "description": "Structure standard pour les réponses réussies",
                "properties": {
                    "success": {
                        "type": "boolean",
                        "example": true
                    },
                    "message": {
                        "type": "string",
                        "example": "Opération réussie"
                    },
                    "data": {
                        "type": "object",
                        "nullable": true
                    },
                    "meta": {
                        "type": "object",
                        "nullable": true
                    }
                },
                "type": "object"
            },
            "ErrorResponse": {
                "title": "Réponse d'erreur standard",
                "description": "Structure standard pour les réponses d'erreur",
                "properties": {
                    "success": {
                        "type": "boolean",
                        "example": false
                    },
                    "message": {
                        "type": "string",
                        "example": "Une erreur est survenue"
                    },
                    "code": {
                        "type": "integer",
                        "example": 400
                    },
                    "errors": {
                        "type": "object",
                        "nullable": true
                    }
                },
                "type": "object"
            },
            "ValidationError": {
                "title": "Erreur de validation",
                "description": "Structure pour les erreurs de validation",
                "allOf": [
                    {
                        "$ref": "#/components/schemas/ErrorResponse"
                    },
                    {
                        "properties": {
                            "errors": {
                                "type": "object",
                                "example": {
                                    "email": [
                                        "L'email est invalide"
                                    ]
                                }
                            }
                        },
                        "type": "object"
                    }
                ]
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Authentification via JWT Token. Entrez votre token sans le préfixe 'Bearer'",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}